+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 ### Added
 
 - Add `discovery.TargetExecutor` and `ClientExecutor`
+- Add `discovery.ApplicationObserver`, `ApplicationObserverSet` and `ApplicationExecutor`
+- Add `discovery.Inspector`
 
 ### Changed